fix: new line added & locations api call moved to js file
