+++ .babelrc
@@ -14,4 +14,3 @@
       }
     }
   }
-
