+++ pages/services.vue
@@ -18,8 +18,8 @@
   </div>
 </template>
 <script>
-import TServiceListItem from '../components/TServiceListItem.vue'
-import { getServices } from '../airtable.js'
+import TServiceListItem from '../components/services/TServiceListItem.vue'
+import { getServices } from '../utils/airtable.js'
 
 export default {
   name: 'Services',
